Refactor the home page by creating a reusable carousel for our movies featured on the home page.
